# openapi.yaml
components:
  schemas:
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
info:
  title: FastAPI
  version: 0.1.0
openapi: 3.0.2
paths:
  /values:
    get:
      description: Retrieves all values from the WellKnownsConfig class
      operationId: retrieve_all_values_values_get
      responses:
        '200':
          content:
            application/json:
              schema:
                title: Response Retrieve All Values Values Get
                type: object
          description: Successful Response
      summary: retrieve all configured values
  /values/{value_name}:
    get:
      description: "Retrieves the given value from configuration\nArgs:\n    value_name:\
        \ the name of the value to be retrieved\n\nRaises:\n    HTTPException 404\
        \ when the specified value is not configured"
      operationId: retrieve_value_values__value_name__get
      parameters:
      - in: path
        name: value_name
        required: true
        schema:
          title: Value Name
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                title: Response Retrieve Value Values  Value Name  Get
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: retrieve a configured value